ci: add a new step for detecting update of nginx
GitHub Actions detect whether we need to update bundled nginx or not
by comparing the latest version of nginx of upstream and the version
of bundled nginx.

Currently, we confirm the latest version of nginx manually.
However, in this method, we often forget to update nginx that is
bundled with Groonga.